+++ packages/camera/camera/example/ios/RunnerTests/CameraCaptureSessionQueueRaceConditionTests.m
@@ -29,10 +29,11 @@
                     result:^(id _Nullable result) {
                       [disposeExpectation fulfill];
                     }];
-  [camera handleMethodCall:createCall
-                    result:^(id _Nullable result) {
-                      [createExpectation fulfill];
-                    }];
+  [camera createCameraOnSessionQueueWithCreateMethodCall:createCall
+                                                  result:[[FLTThreadSafeFlutterResult alloc]
+                                                             initWithResult:^(id _Nullable result) {
+                                                               [createExpectation fulfill];
+                                                             }]];
   [self waitForExpectationsWithTimeout:1 handler:nil];
   // `captureSessionQueue` must not be nil after `create` call. Otherwise a nil
   // `captureSessionQueue` passed into `AVCaptureVideoDataOutput::setSampleBufferDelegate:queue:`