+++ packages/camera/camera/example/lib/readme_full_example.dart
@@ -36,6 +36,17 @@ class _CameraAppState extends State<CameraApp> {
         return;
       }
       setState(() {});
+    }).catchError((Object e) {
+      if (e is CameraException) {
+        switch (e.code) {
+          case 'CameraAccessDenied':
+            print('User denied camera access.');
+            break;
+          default:
+            print('Handle other errors.');
+            break;
+        }
+      }
     });
   }