+++ packages/path_provider/path_provider_windows/lib/src/path_provider_windows_real.dart
@@ -13,21 +13,43 @@ import 'package:win32/win32.dart';
 
 import 'folders.dart';
 
+/// Constant for en-US language used in VersionInfo keys.
+@visibleForTesting
+const String languageEn = '0409';
+
+/// Constant for CP1252 encoding used in VersionInfo keys
+@visibleForTesting
+const String encodingCP1252 = '04e4';
+
+/// Constant for Unicode encoding used in VersionInfo keys
+@visibleForTesting
+const String encodingUnicode = '04b0';
+
 /// Wraps the Win32 VerQueryValue API call.
 ///
 /// This class exists to allow injecting alternate metadata in tests without
 /// building multiple custom test binaries.
 @visibleForTesting
 class VersionInfoQuerier {
-  /// Returns the value for [key] in [versionInfo]s English strings section, or
-  /// null if there is no such entry, or if versionInfo is null.
-  String? getStringValue(Pointer<Uint8>? versionInfo, String key) {
+  /// Returns the value for [key] in [versionInfo]s in section with given
+  /// language and encoding, or null if there is no such entry,
+  /// or if versionInfo is null.
+  ///
+  /// See https://docs.microsoft.com/en-us/windows/win32/menurc/versioninfo-resource
+  /// for list of possible language and encoding values.
+  String? getStringValue(
+    Pointer<Uint8>? versionInfo,
+    String key, {
+    required String language,
+    required String encoding,
+  }) {
+    assert(language.isNotEmpty);
+    assert(encoding.isNotEmpty);
     if (versionInfo == null) {
       return null;
     }
-    const String kEnUsLanguageCode = '040904e4';
     final Pointer<Utf16> keyPath =
-        TEXT('\\StringFileInfo\\$kEnUsLanguageCode\\$key');
+        TEXT('\\StringFileInfo\\$language$encoding\\$key');
     final Pointer<Uint32> length = calloc<Uint32>();
     final Pointer<Pointer<Utf16>> valueAddress = calloc<Pointer<Utf16>>();
     try {
@@ -150,6 +172,12 @@ class PathProviderWindows extends PathProviderPlatform {
     }
   }
 
+  String? _getStringValue(Pointer<Uint8>? infoBuffer, String key) =>
+      versionInfoQuerier.getStringValue(infoBuffer, key,
+          language: languageEn, encoding: encodingCP1252) ??
+      versionInfoQuerier.getStringValue(infoBuffer, key,
+          language: languageEn, encoding: encodingUnicode);
+
   /// Returns the relative path string to append to the root directory returned
   /// by Win32 APIs for application storage (such as RoamingAppDir) to get a
   /// directory that is unique to the application.
@@ -187,10 +215,10 @@ class PathProviderWindows extends PathProviderPlatform {
           infoBuffer = null;
         }
       }
-      companyName = _sanitizedDirectoryName(
-          versionInfoQuerier.getStringValue(infoBuffer, 'CompanyName'));
-      productName = _sanitizedDirectoryName(
-          versionInfoQuerier.getStringValue(infoBuffer, 'ProductName'));
+      companyName =
+          _sanitizedDirectoryName(_getStringValue(infoBuffer, 'CompanyName'));
+      productName =
+          _sanitizedDirectoryName(_getStringValue(infoBuffer, 'ProductName'));
 
       // If there was no product name, use the executable name.
       productName ??=