+++ packages/path_provider/path_provider_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Minor fixes for new analysis options.
+
 ## 2.0.6
 
 * Fixes library_private_types_in_public_api, sort_child_properties_last and use_key_in_widget_constructors